Move ConcatSource into its own module.
Also added some more tests.
